Lecture 9/requirements.txt
torchvision==0.5.0
scipy==1.4.1
scikit_image==0.16.2
torch==1.7.0
pandas==1.1.2
numpy==1.19.2
matplotlib==3.3.1
Pillow==8.1.0
scikit_learn==0.24.1
